craftysane/requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
bs4==0.0.1
chardet==3.0.4
colorama==0.4.4
configparser==5.0.0
Django==2.1.2
django-admin-thumbnails==0.2.5
django-ckeditor==5.9.0
django-mptt==0.11.0
django-timezone-field==4.0
docutils==0.15.2
idna==2.9
isort==4.3.21
jsonfield==2.1.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
payu-biz==1.2.1
paywix==1.3.0
Pillow==7.1.2
pyasn1==0.4.8
PyJWT==1.7.1
pylint==2.5.2
python-gettext==4.0
python-dateutil==2.8.1
pytz==2020.1
psycopg2==2.8.6
PyYAML==5.3.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.2
urllib3==1.25.9
whitenoise==5.0.1
wrapt==1.12.1
django-storages==1.11.1
boto3==1.16.63
botocore==1.19.63
awsebcli==3.19.3
s3transfer==0.3.4
jmespath==0.10.0